pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag"
version = "0.1.0"
authors = [
    {name = "Mila Rao", email = "milasrao@gmail.com"},
]
requires-python = "<3.13,>=3.10"
dependencies = [
    "unstructured>=0.10.30",
    "chonkie[semantic,st]>=0.1.2",
    "polars>=0.19.3",
    "chromadb>=0.4.18",
    "crewai>=0.108.0",
    "openai>=1.3.0",
    "streamlit>=1.29.0",
    "unstructured-inference>=0.7.12",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "python-docx>=0.8.11",
    "openpyxl>=3.1.2",
    "tabula-py>=2.8.2",
    "beautifulsoup4>=4.12.2",
    "rank-bm25>=0.2.2",
    "nltk>=3.8.1",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.1",
    "loguru>=0.7.0",
    "pytest>=7.4.3",
    "torch==2.1.0",
    "numpy<2",
    "pi-heif>=0.20.0",
    "torchvision==0.16.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3"
]
